+++ crates/adkg/src/aba/crain20.rs
@@ -1,31 +1,32 @@
 //! Implementation of the Tyler Crain's Asynchronous Byzantine Agreement described in https://arxiv.org/pdf/2002.08765.
 //! We specifically implement the Good-Case-Coin-Free variant described in https://eprint.iacr.org/2021/1591.pdf, Appendix B.
 
-use futures::StreamExt;
+mod broadcast;
+mod coin;
 mod ecdh_coin_toss;
 pub mod messages;
+mod recv_handler;
 
 use crate::aba::{Aba, AbaConfig, Estimate};
 use crate::helpers::{PartyId, SessionId};
 use crate::network::{RetryStrategy, broadcast_with_self};
 use ark_ec::CurveGroup;
 use dcipher_network::topic::TopicBasedTransport;
-use dcipher_network::{ReceivedMessage, Transport, TransportSender};
+use dcipher_network::{Transport, TransportSender};
 use digest::core_api::BlockSizeUser;
 use digest::crypto_common::rand_core::CryptoRng;
 use digest::{DynDigest, FixedOutputReset};
-use ecdh_coin_toss::{Coin, EcdhCoinTossError, EcdhCoinTossEval};
+use ecdh_coin_toss::{EcdhCoinTossError, EcdhCoinTossEval};
 use futures::future::Either;
-use messages::{AbaMessage, AuxStage, CoinEvalMessage, View};
-use messages::{AuxiliaryMessage, AuxiliarySetMessage, EstimateMessage};
+use messages::AuxiliarySetMessage;
+use messages::{AbaMessage, AuxStage, View};
 use rand::RngCore;
-use serde::{Deserialize, Serialize};
-use std::collections::{BTreeSet, HashSet};
+use serde::Deserialize;
+use std::collections::BTreeSet;
 use std::hash::Hash;
 use std::ops::{Deref, DerefMut, Index, IndexMut};
 use std::pin::pin;
 use std::{
-    borrow::Borrow,
     collections::{BTreeMap, HashMap, btree_map::Entry},
     marker::PhantomData,
     sync::Arc,
@@ -35,7 +36,7 @@ use tokio::sync::oneshot::error::RecvError;
 use tokio::sync::{Mutex, Notify, oneshot};
 use tokio::task::JoinError;
 use tokio_util::sync::CancellationToken;
-use tracing::{Instrument, Level, debug, error, event, info, trace, warn};
+use tracing::{Level, debug, error, event, info};
 use utils::hash_to_curve::HashToCurve;
 use utils::serialize::fq::FqSerialize;
 use utils::serialize::point::PointSerializeCompressed;
@@ -180,10 +181,7 @@ where
     receiver: T::ReceiveMessageStream,
 }
 
-struct AbaCrain20Instance<CG, CK, H, TS>
-where
-    TS: TransportSender,
-{
+struct AbaCrain20Instance<CG, CK, H, TS> {
     config: Arc<AbaCrain20Config<CG, CK, H>>,
     sid: SessionId,
     sender: TS,
@@ -300,171 +298,6 @@ where
                 error!("Node `{}` failed to abort recv thread: {e:?}.", config.id);
                 Err(AbaError::Join(e, "failed to join recv thread").into())
             }
-        }
-    }
-}
-
-impl<CG, CK, H, T> AbaCrain20<CG, CK, H, T>
-where
-    CG: CurveGroup + Copy + HashToCurve + PointSerializeCompressed,
-    CG::ScalarField: FqSerialize,
-    EcdhCoinTossEval<CG, H>: for<'de> Deserialize<'de>,
-    CK: Send + Into<CoinKeys<CG>> + 'static,
-    H: Default + DynDigest + BlockSizeUser + Clone + Send + Sync + 'static,
-    T: Transport<Identity = PartyId>,
-    T::Sender: Clone,
-{
-    /// Thread responsible for receiving all types of ABA messages and transmitting notifications.
-    async fn recv_thread(
-        sid: SessionId,
-        config: Arc<AbaCrain20Config<CG, CK, H>>,
-        receiver: T::ReceiveMessageStream,
-        sender: T::Sender,
-        cancel: CancellationToken,
-        state: Arc<AbaState<CG, H>>,
-    ) {
-        let id = config.id;
-        // Stop the thread upon receiving a signal from the cancellation token
-        tokio::select! {
-            _ = cancel.cancelled() => {
-                info!("Node `{id}` in ABA with sid `{sid}` stopping recv_thread");
-            }
-
-            _ = Self::recv_loop(config, receiver, sender, state).instrument(tracing::info_span!("recv_loop", ?sid)) => {}
-        }
-    }
-
-    /// Infinite loop listening for ABA messages and sending notifications.
-    async fn recv_loop(
-        config: Arc<AbaCrain20Config<CG, CK, H>>,
-        mut receiver: T::ReceiveMessageStream,
-        sender: T::Sender,
-        state: Arc<AbaState<CG, H>>,
-    ) {
-        // Local variables
-        let mut count_est = PerPartyStorage::new();
-        let mut sent_estimate: HashSet<EstimateMessage> = HashSet::new();
-
-        loop {
-            let ReceivedMessage {
-                sender: sender_id,
-                content,
-                ..
-            } = match receiver.next().await {
-                Some(Ok(m)) => m,
-                Some(Err(e)) => {
-                    warn!("Node `{}` failed to recv: {e:?}", config.id);
-                    continue;
-                }
-                None => {
-                    error!(
-                        "Node `{}` failed to recv: no more items in stream",
-                        config.id
-                    );
-                    return;
-                }
-            };
-
-            let m: AbaMessage = match bson::from_slice(&content) {
-                Ok(m) => m,
-                Err(e) => {
-                    error!(error = ?e, "Node `{}` failed to deserialize message", config.id);
-                    continue;
-                }
-            };
-            trace!(
-                "Node `{}` received message {m:?} from {sender_id}",
-                config.id,
-            );
-
-            match m {
-                // 4: upon receiving BVAL(v) do
-                AbaMessage::Estimate(est) => {
-                    count_est.insert_once(est, sender_id, true);
-                    let count = count_est.get_count(&est);
-
-                    #[allow(clippy::int_plus_one)]
-                    if count >= 2 * config.t + 1 {
-                        // 7: if BVAL(V) received from 2t + 1 different nodes then
-                        // 8: bin_values := bin_values \cup {v}
-                        // add the estimate to the binary values
-                        let mut r_bin_values = state.bin_values.lock().await;
-                        let bin_values = &mut r_bin_values.entry(est.round).or_default()[est.stage];
-                        if bin_values.contains(&est.estimate) {
-                            drop(r_bin_values);
-                        } else {
-                            bin_values.push(est.estimate);
-                            drop(r_bin_values);
-
-                            // notify of update to bin_values
-                            debug!(
-                                "Node {} notifying bin values for round {}",
-                                config.id, est.round
-                            );
-                            state.notify_bin_values.notify_one((est.round, est.stage));
-                        };
-                    } else if count >= config.t + 1 && !sent_estimate.contains(&est) {
-                        // 5: if BVAL(v) received from t + 1 different nodes AND BVAL(v) was not sent, then
-                        // 6: Send BVAL(v) to all nodes
-                        let msg_est = AbaMessage::Estimate(est);
-                        if let Err(e) =
-                            broadcast_with_self(&msg_est, &config.retry_strategy, &sender).await
-                        {
-                            error!(
-                                "Node `{}` failed to broadcast estimate message: {e:?}",
-                                config.id
-                            )
-                        }
-                        sent_estimate.insert(est);
-                    }
-                }
-
-                AbaMessage::Auxiliary(aux) => {
-                    let mut aux_views = state.aux_views.lock().await;
-                    // Add the new estimate to the current view
-                    aux_views
-                        .entry((aux.round, aux.stage), sender_id)
-                        .or_default()
-                        .insert(aux.estimate);
-
-                    // notify once we got at least n - t aux messages
-                    if aux_views.get_count(&(aux.round, aux.stage)) >= config.n - config.t {
-                        state.notify_count_aux.notify_one((aux.round, aux.stage));
-                    }
-                }
-
-                AbaMessage::AuxiliarySet(aux_set) => {
-                    // Insert auxset view, at most once per sender_id
-                    let mut auxset_views = state.auxset_views.lock().await;
-                    auxset_views.insert_once(aux_set.round, sender_id, aux_set.view);
-
-                    // notify once we got at least n - t auxset messages
-                    if auxset_views.get_count(&aux_set.round) >= config.n - config.t {
-                        state.notify_count_auxset.notify_one(aux_set.round);
-                    }
-                }
-
-                AbaMessage::CoinEval(msg_eval) => {
-                    // Deserialize eval
-                    let Ok(eval): Result<EcdhCoinTossEval<CG, _>, _> = msg_eval.borrow().try_into()
-                    else {
-                        warn!("Failed to deserialize CoinEvalMessage");
-                        continue;
-                    };
-
-                    // Store one eval per party, per round. We cannot verify it here
-                    // since the node may not be ready to check evaluations yet.
-                    let mut coin_evals = state.coin_evals.lock().await;
-                    coin_evals.insert_once(msg_eval.round, sender_id, eval);
-                    let count = coin_evals.get_count(&msg_eval.round);
-                    drop(coin_evals); // drop lock
-
-                    // Notify if we have t + 1 evals
-                    if count > config.t {
-                        state.notify_enough_coin_evals.notify_one(msg_eval.round);
-                    }
-                }
-            };
         }
     }
 }
@@ -638,291 +471,9 @@ where
             }
         }
     }
+}
 
-    /// Binary-value broadcast described in https://dl.acm.org/doi/10.1145/2785953, Figure 1
-    /// Send the current party's estimate to all other nodes with an Estimate message.
-    #[tracing::instrument(skip(self))]
-    async fn bv_broadcast(&self, r: u8, stage: AuxStage, v: Estimate) {
-        // 1: broadcast B_VAL(v) to all
-        let msg_est = AbaMessage::Estimate(EstimateMessage {
-            round: r,
-            stage,
-            estimate: v,
-        });
-
-        event!(
-            Level::DEBUG,
-            "Node `{}` at round `{r}` sending {:?} to all",
-            self.config.id,
-            msg_est
-        );
-        if let Err(e) =
-            broadcast_with_self(&msg_est, &self.config.retry_strategy, &self.sender).await
-        {
-            error!(
-                "Node `{}` failed to broadcast estimate message: {e:?}",
-                self.config.id
-            )
-        }
-    }
-
-    /// Synchronized binary-value broadcast described in https://dl.acm.org/doi/10.1145/2785953, Figure 2
-    /// Send the current party's estimate to all other nodes with an Estimate message.
-    #[tracing::instrument(skip(self, state))]
-    async fn sbv_broadcast(
-        &self,
-        r: u8,
-        stage: AuxStage,
-        v: Estimate,
-        state: &Arc<AbaState<CG, H>>,
-    ) -> View {
-        // 1: BV_Broadcast(v)
-        self.bv_broadcast(r, stage, v).await;
-
-        event!(
-            Level::DEBUG,
-            "Node `{}` waiting for bin values",
-            self.config.id
-        );
-        let bin_values = loop {
-            // 2: wait until bin_values \neq \emptyset
-            state.notify_bin_values.notified((r, stage)).await;
-
-            let bin_values = state.bin_values.lock().await;
-            let bin_values = &bin_values.get(&r).cloned().unwrap_or_default()[stage];
-            if !bin_values.is_empty() {
-                event!(
-                    Level::DEBUG,
-                    "Node `{}` obtained bin_values = `{bin_values:?}`",
-                    self.config.id
-                );
-                break bin_values.clone();
-            }
-        };
-
-        // 3: Send AUX(w) for w \in bin_values to all
-        for w in bin_values.iter() {
-            let msg_aux = AbaMessage::Auxiliary(AuxiliaryMessage {
-                round: r,
-                stage,
-                estimate: *w,
-            });
-            event!(
-                Level::DEBUG,
-                "Node `{}` sending {:?} to all",
-                self.config.id,
-                msg_aux
-            );
-
-            if let Err(e) =
-                broadcast_with_self(&msg_aux, &self.config.retry_strategy, &self.sender).await
-            {
-                error!(
-                    "Node `{}` failed to broadcast aux message: {e:?}",
-                    self.config.id
-                )
-            }
-        }
-
-        // 4: wait until \exists a set view s.t.
-        //  (1) view \subseteq bin_values, and
-        //  (2) contained in AUX(.) messages received from n - t nodes
-        let view = loop {
-            event!(
-                Level::DEBUG,
-                "Node `{}` waiting for count_aux notification",
-                self.config.id
-            );
-
-            // wake up each time after having received n - t aux, or on bin_values update
-            future_select_pin(
-                state.notify_count_aux.notified((r, stage)),
-                state.notify_bin_values.notified((r, stage)),
-            )
-            .await;
-
-            let aux_views = state.aux_views.lock().await;
-            let bin_values = state.bin_values.lock().await; // warn: two locks, could deadlock
-            let aux_views = aux_views.get(&(r, stage)).to_owned().unwrap_or_default();
-            let bin_values = &bin_values.get(&r).cloned().unwrap_or_default()[stage];
-            let view = self.construct_view(bin_values, &aux_views);
-            if let Some(view) = view {
-                event!(
-                    Level::DEBUG,
-                    "Node {} obtained view = `{view:?}`",
-                    self.config.id
-                );
-                break view;
-            } else {
-                event!(
-                    Level::DEBUG,
-                    "Node {} received notify_count_aux notification while having no binary estimates / not enough aux",
-                    self.config.id
-                );
-            }
-        };
-        // 5: return view
-        #[allow(clippy::let_and_return)] // for clarity
-        view
-    }
-
-    /// Try to get the output from the coin keys receiver, return an error otherwise.
-    async fn get_coin_keys(
-        &self,
-        r: u8,
-        coin_keys_receiver: oneshot::Receiver<CK>,
-    ) -> Result<CK, AbaError> {
-        event!(
-            Level::DEBUG,
-            "Node `{}` at round `{r}` has not yet obtained keys for common coin protocol, waiting.",
-            self.config.id
-        );
-
-        // Return coin_keys if sender not dropped, err otherwise
-        match coin_keys_receiver.await {
-            Ok(coin_keys) => {
-                event!(
-                    Level::DEBUG,
-                    "Node `{}` at round `{r}` obtained keys for common coin protocol",
-                    self.config.id
-                );
-
-                Ok(coin_keys)
-            }
-            Err(_) => {
-                error!(
-                    "Node `{}` at round `{r}` failed to obtain common coin input through channel: sender dropper. Aborting ABA.",
-                    self.config.id
-                );
-                Err(AbaError::CoinKeysRecv)
-            }
-        }
-    }
-
-    /// Try to generate and send a partial coin evaluation, or return an error otherwise.
-    async fn send_coin_eval<RNG>(
-        &self,
-        r: u8,
-        coin_keys: &CoinKeys<CG>,
-        rng: &mut RNG,
-    ) -> Result<(), Box<AbaError>>
-    where
-        RNG: RngCore + CryptoRng,
-    {
-        let eval = EcdhCoinTossEval::<CG, H>::eval(
-            &coin_keys.sk,
-            &Self::coin_input(usize::from(self.sid), &coin_keys.combined_vk, r)?,
-            &self.config.g,
-            rng,
-        )
-        .map_err(|e| AbaError::CoinToss(e, "failed to generate coin toss evaluation: {e}"))?;
-
-        let msg_coin_eval = AbaMessage::CoinEval(CoinEvalMessage::new(eval, r).unwrap());
-
-        if let Err(e) =
-            broadcast_with_self(&msg_coin_eval, &self.config.retry_strategy, &self.sender).await
-        {
-            error!(
-                "Node `{}` failed to broadcast coin eval message: {e:?}",
-                self.config.id
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Wait for enough evaluations and try to recover a common coin. Returns an error if too many evaluations are invalid.
-    async fn get_coin(
-        &self,
-        r: u8,
-        state: &Arc<AbaState<CG, H>>,
-        coin_keys: &CoinKeys<CG>,
-    ) -> Result<Coin, Box<AbaError>> {
-        // Get the input of the common coin protocol
-        let coin_input = Self::coin_input(
-            usize::from(self.sid),
-            &coin_keys.combined_vk.into_affine().into(),
-            r,
-        )?;
-
-        loop {
-            // Wait until we have enough valid partial coins evals for the current round
-            event!(
-                Level::DEBUG,
-                "Node `{}` at round `{r}` waiting for coin evaluations",
-                self.config.id
-            );
-            state.notify_enough_coin_evals.notified(r).await;
-
-            // mutex locked for the entire duration, either that or cloning evals
-            let coin_evals = state.coin_evals.lock().await;
-            let Some((senders, evals)) = coin_evals.get_all(&r) else {
-                event!(
-                    Level::DEBUG,
-                    "Node `{}` at round `{r}` received coin evals notifications while not having evals",
-                    self.config.id
-                );
-                continue;
-            };
-
-            if evals.len() < self.config.t + 1 {
-                event!(
-                    Level::DEBUG,
-                    "Node `{}` at round `{r}` does not have enough evals: {} < {}",
-                    self.config.id,
-                    evals.len(),
-                    self.config.t
-                );
-                continue; // not enough evals for this round yet
-            };
-
-            // Try to get and return the common coin
-            let coin_vks: Vec<_> = senders.iter().map(|&j| coin_keys.vks[j]).collect();
-            match EcdhCoinTossEval::get_coin(
-                &evals,
-                &senders,
-                &coin_vks,
-                &coin_input,
-                &self.config.g,
-                self.config.t + 1,
-            ) {
-                Ok(coin) => return Ok(coin),
-                Err(e) => {
-                    // Failed to obtain the common coin, we either continue if we don't have all evals yet, or we abort
-                    event!(
-                        Level::WARN,
-                        "Node `{}` at round `{r}` failed to obtain a common coin due to invalid eval(s): {e:?}",
-                        self.config.id
-                    );
-
-                    if evals.len() < self.config.n {
-                        continue;
-                    } else {
-                        event!(
-                            Level::ERROR,
-                            "Node `{}` at round `{r}` failed to obtain a common coin with n evals: {e:?}. Aborting ABA with error.",
-                            self.config.id
-                        );
-                        Err(AbaError::CoinToss(
-                            e,
-                            "failed to obtain common coin with all evals",
-                        ))?
-                    }
-                }
-            }
-        }
-    }
-
-    /// Get the input to the common coin.
-    fn coin_input(sid: usize, combined_vk: &CG, round: u8) -> Result<Vec<u8>, Box<AbaError>> {
-        CoinInput {
-            combined_vk: *combined_vk,
-            sid,
-            round,
-        }
-        .serialize()
-    }
-
+impl<CG, CK, H, TS> AbaCrain20Instance<CG, CK, H, TS> {
     /// Try to build a view from the union of views sent by other nodes, filtered by local binary values
     /// obtained through the BV_broadcast algorithm, Figure 1 of <https://arxiv.org/pdf/2002.08765>.
     /// Implements filtering of line (05), Figure 3 of <https://arxiv.org/pdf/2002.08765>:
@@ -951,26 +502,6 @@ where
         }
 
         None
-    }
-}
-
-/// Structure used to serialize the input of the coin
-#[derive(Serialize)]
-#[serde(bound(serialize = "CG: PointSerializeCompressed",))]
-struct CoinInput<CG> {
-    #[serde(with = "utils::serialize::point::base64")]
-    combined_vk: CG,
-    sid: usize,
-    round: u8,
-}
-
-impl<CG> CoinInput<CG>
-where
-    CG: PointSerializeCompressed,
-{
-    fn serialize(&self) -> Result<Vec<u8>, Box<AbaError>> {
-        bson::to_vec(&self)
-            .map_err(|e| AbaError::BsonSer(e, "failed to serialize CoinInput to bson").into())
     }
 }
